Combine home and notification page and store ID in local storage
