+++ package.json
@@ -29,7 +29,7 @@
     "moment": "^2.17.1",
     "qrcode.react": "^0.6.1",
     "react": "^0.13.0",
-    "react-router": "^0.13.3",
+    "react-router": "^2.0.0",
     "shortid": "^2.2.6",
     "socket.io": "^1.7.2",
     "socket.io-client": "^1.7.2"
